# Conflicts:
#	.idea/workspace.xml
#	src/main/java/org/tienda/Controller/cForgotPasswordMail.java
#	src/main/java/org/tienda/Controller/controllerLogin.java
#	src/main/java/org/tienda/Controller/controllerRegister.java
#	src/main/java/org/tienda/Models/modelForgotPasswordPassword.java
#	src/main/java/org/tienda/Objects/Historialusuarios.java
#	src/main/java/org/tienda/Utils/EmailUtil.java
#	src/main/java/org/tienda/Views/ForgotPasswordPassword.form
#	src/main/java/org/tienda/Views/ForgotPasswordPassword.java
#	src/main/resources/hibernate.cfg.xml
#	src/main/resources/modelo/Carrito.hbm.xml
#	src/main/resources/modelo/Historialusuarios.hbm.xml
#	src/main/resources/modelo/Usuarios.hbm.xml
#	tiendanavidad.iml

+++ src/main/resources/sql/estructura.sql
@@ -2,13 +2,13 @@ CREATE DATABASE IF NOT EXISTS tienda;
 
 CREATE TABLE IF NOT EXISTS productos (
   idproducto  INT AUTO_INCREMENT PRIMARY KEY,
-  nombre      VARCHAR(50)   NOT NULL,
-  descripcion VARCHAR(255)  NOT NULL,
-  precio      DOUBLE(10, 2) NOT NULL,
-  descuento   DOUBLE(3, 2)  NULL,
-  categoria   VARCHAR(30)   NULL,
-  imagen      LONGBLOB      NULL,
-  stock       INT           NULL,
+  nombre      VARCHAR(50)  NOT NULL,
+  descripcion VARCHAR(255) NOT NULL,
+  precio      DOUBLE       NOT NULL,
+  descuento   DOUBLE       NULL,
+  categoria   VARCHAR(30)  NULL,
+  imagen      LONGBLOB     NULL,
+  stock       INT          NULL,
   CHECK (precio > 0)
 );
